% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydscompanion.R
\docType{package}
\name{tidydscompanion}
\alias{tidydscompanion}
\alias{tidydscompanion-package}
\title{tidydscompanion: A companion package for Using R and the tidyverse for Data
Science}
\description{
Provides data sets and installation help for the Using R and the tidyverse
for Data Science workshop
}
